Consolidated all the console app code into the console project and removed unnecessary using statements.
