+++ ExchangeSharpConsole_Tests.cs
@@ -12,14 +12,12 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
 
 using System;
 using System.Collections.Generic;
-using System.IO;
 using System.Linq;
-using System.Runtime.CompilerServices;
-using System.Threading;
+using ExchangeSharp;
 
-namespace ExchangeSharp
+namespace ExchangeSharpConsoleApp
 {
-    public static partial class ExchangeSharpConsole
+	public static partial class ExchangeSharpConsole
     {
         private static void Assert(bool expression)
         {
